rename URL and navigation. Issue #20
